Refactor input checking into a common function in utils and create separate logging and matrix ops files to remove circular dependencies
