+++ rec/__init__.py
@@ -1,3 +1,3 @@
 """ T-RECS simulation package """
 from rec import models, components, metrics, random
-from .utils import VerboseMode
+from .logging import VerboseMode
