+++ rec/components/users.py
@@ -5,7 +5,7 @@ encapsulates some of these concepts)
 """
 import numpy as np
 
-from rec.utils import contains_row, slerp
+from rec.matrix_ops import contains_row, slerp
 from rec.random import Generator
 from .base_components import Component, BaseComponent
