+++ rec/metrics/measurement.py
@@ -6,7 +6,7 @@ from abc import ABC, abstractmethod
 import networkx as nx
 from networkx import wiener_index
 import numpy as np
-from rec.utils import VerboseMode
+from rec.logging import VerboseMode
 from rec.components import (
     BaseObservable,
     register_observables,
